CapstoneP_PhaseA_Core_Agentic_Engine_MVP3_DB/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
httpx==0.25.2
python-dotenv==1.0.0
openai==1.12.0
duckduckgo-search==6.1.1
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0

